Java

+++ src/Lesson4/Frame5/Notebook.java
@@ -3,7 +3,7 @@ package Lesson4.Frame5;
 public class Notebook {
     private String brend;
     private String model;
-    private int serialNumber;
+    private long serialNumber;
     private boolean wiFi;
     private boolean blutooth;
 }
